lambdas/image-processor/requirements.txt
boto3==1.28.62
pillow==10.0.1